python/requirements.txt
websocket-client
aiohttp
flask
flask_sqlalchemy
flask_migrate